move all EIR fitting functions to AnophelesModelFitter.h
